pyproject.toml
[project]
name = "spp-weis-price-forecast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3==1.35.92",
    "darts==0.31.0",
    "duckdb==1.3.2",
    "jupyterlab>=4.5.2",
    "lightning>=2.6.0",
    "lightning-sdk>=2025.12.17",
    "meteostat>=2.0.1",
    "mlflow==2.16.0",
    "optuna-dashboard>=0.20.0",
    "optuna-integration[pytorch-lightning]>=4.6.0",
    "pandas>=2.3.3",
    "plotly>=6.0",
    "polars>=1.37.1",
    "polars-xdt>=0.17.1",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.2.1",
    "rsconnect-python>=1.28.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "shiny>=1.5.1",
    "shinyswatch>=0.9.0",
    "shinywidgets>=0.7.1",
    "streamlit==1.38",
    "torch<2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[dependency-groups]
dev = [
    "marimo>=0.13.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
]
